requirements.txt
requests
python-dotenv
openai==1.13.3
httpx==0.27.0
prometheus_client
flask